requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
python-decouple==3.8
mysqlclient==2.1.1
django-filter==23.3